新增修改用户功能

+++ off-work-services/off-work-service-user/src/main/java/com/nicr0n/user/entity/po/SysUserUpdateDTO.java
@@ -54,5 +54,5 @@ public class SysUserUpdateDTO {
     private LocalDate birthday;
 
     @ApiModelProperty("角色ID列表")
-    private List<Integer> roleIDList;
+    private List<Long> roleIDList;
 }
